requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
channels==4.0.0
charset-normalizer==3.3.2
Django==4.2.7
django-admin-interface==0.26.1
django-colorfield==0.10.1
django-cors-headers==4.3.0
django-nextjs==2.4.0
djangorestframework==3.14.0
Faker==20.0.3
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-environ==0.4.54
python-slugify==8.0.1
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
urllib3==2.0.7
websockets==12.0
yarl==1.9.2
